+++ cypress/elements/dashboardItem.js
@@ -1,3 +1,5 @@
+import { EXTENDED_TIMEOUT } from '../support/utils'
+
 export const chartSel = '.highcharts-container'
 export const chartSubtitleSel = '.highcharts-subtitle'
 export const chartXAxisLabelSel = '.highcharts-xaxis-labels'
@@ -6,13 +8,19 @@ export const tableSel = '.pivot-table-container'
 export const gridItemSel = '.react-grid-item'
 
 export const itemDetailsSel = '[data-test="dashboarditem-footer"]'
-export const itemMenuButton = '[data-test="dashboarditem-menu-button"]'
+export const itemMenuButtonSel = '[data-test="dashboarditem-menu-button"]'
 
-export const getDashboardItem = name =>
-    cy.get(`[data-test="dashboarditem-${name}"]`)
+export const getDashboardItem = itemUid =>
+    cy.get(`[data-test="dashboarditem-${itemUid}"]`, EXTENDED_TIMEOUT)
 
-export const clickMenuButton = name =>
-    getDashboardItem(name).find(itemMenuButton).click()
+export const clickMenuButton = itemUid =>
+    getDashboardItem(itemUid).scrollIntoView().find(itemMenuButtonSel).click()
+
+export const clickItemDeleteButton = itemUid =>
+    getDashboardItem(itemUid)
+        .scrollIntoView()
+        .find('[data-test="delete-item-button"]')
+        .click()
 
 //map