+++ src/components/Item/VisualizationItem/Item.js
@@ -219,25 +219,31 @@ class Item extends Component {
                     )}
                     onFatalError={this.onFatalError}
                 >
-                    <div
-                        className="dashboard-item-content"
-                        ref={ref => (this.contentRef = ref)}
-                    >
+                    <div ref={ref => (this.contentRef = ref)}>
                         {this.state.configLoaded && (
                             <WindowDimensionsCtx.Consumer>
                                 {dimensions => (
-                                    <Visualization
-                                        item={item}
-                                        activeType={activeType}
-                                        itemFilters={itemFilters}
-                                        availableHeight={this.getAvailableHeight(
-                                            dimensions
-                                        )}
-                                        availableWidth={this.getAvailableWidth()}
-                                        gridWidth={this.props.gridWidth}
-                                        dashboardMode={dashboardMode}
-                                        isRecording={this.props.isRecording}
-                                    />
+                                    <div
+                                        className="dashboard-item-content"
+                                        style={{
+                                            height: this.getAvailableHeight(
+                                                dimensions
+                                            ),
+                                        }}
+                                    >
+                                        <Visualization
+                                            item={item}
+                                            activeType={activeType}
+                                            itemFilters={itemFilters}
+                                            availableHeight={this.getAvailableHeight(
+                                                dimensions
+                                            )}
+                                            availableWidth={this.getAvailableWidth()}
+                                            gridWidth={this.props.gridWidth}
+                                            dashboardMode={dashboardMode}
+                                            isRecording={this.props.isRecording}
+                                        />
+                                    </div>
                                 )}
                             </WindowDimensionsCtx.Consumer>
                         )}