+++ src/components/Item/VisualizationItem/Visualization/DefaultPlugin.js
@@ -60,6 +60,7 @@ const DefaultPlugin = ({
             load(item, visualization, {
                 credentials,
                 activeType,
+                options,
             })
         }