+++ src/modules/itemTypes.js
@@ -35,7 +35,7 @@ const DOMAIN_TYPE_TRACKER = 'TRACKER'
 
 // Dashboard helpers
 export const isVisualizationType = item =>
-    !!itemTypeMap[item.type].isVisualizationType
+    !!itemTypeMap[item.type]?.isVisualizationType
 export const hasMapView = itemType =>
     itemTypeMap[itemType].domainType === DOMAIN_TYPE_AGGREGATE
 export const isTrackerDomainType = itemType =>
@@ -51,6 +51,7 @@ export const itemTypeMap = {
         endPointName: 'visualizations',
         propName: 'visualization',
         pluralTitle: i18n.t('Visualizations'),
+        isVisualizationType: true,
         appUrl: id => `dhis-web-data-visualizer/#/${id}`,
         appName: 'Data Visualizer',
         defaultItemCount: 10,